+++ tests/testthat/test_add_gurobi_solver.R
@@ -22,6 +22,45 @@ test_that("format", {
   expect_gt(ncol(s), 0)
 })
 
+test_that("linear objective", {
+  # define skips
+  skip_on_cran()
+  skip_if_not_installed("gurobi", "13.0.0")
+  # create data
+  projects <- tibble::tibble(
+    name = c("P1", "P2", "P3", "P4"),
+    success = c(0.95, 0.96, 0.94, 1.00),
+    F1 = c(0.91, 0.00, 0.80, 0.10),
+    F2 = c(0.00, 0.92, 0.80, 0.10),
+    F3 = c(0.00, 0.00, 0.00, 0.10),
+    A1 = c(TRUE, FALSE, FALSE, FALSE),
+    A2 = c(FALSE, TRUE, FALSE, FALSE),
+    A3 = c(FALSE, FALSE, TRUE, FALSE),
+    A4 = c(FALSE, FALSE, FALSE, TRUE)
+  )
+  actions <- tibble::tibble(
+    name = c("A1", "A2", "A3", "A4"),
+    cost = c(0.10, 0.10, 0.15, 0)
+  )
+  features <- tibble::tibble(name = c("F1", "F2", "F3"))
+  # build problem
+  p <-
+    problem(
+      projects, actions, features, "name", "success", "name", "cost",
+      "name"
+    ) %>%
+    add_max_wtd_sum_objective(budget = 0.16) %>%
+    add_binary_decisions() %>%
+    add_gurobi_solver()
+  # solve problem
+  s <- solve(p)
+  # run tests
+  expect_equal(s$A1, 0)
+  expect_equal(s$A2, 0)
+  expect_equal(s$A3, 1)
+  expect_equal(s$A4, 1)
+})
+
 test_that("multiple solutions (single objective)", {
   # define skips
   skip_on_cran()